requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
python-dotenv
Flask-Login
